refactor: Organize files into folders
- 기능 수정 없음
- 파일을 종류 별로 폴더에 정리함
