+++ app/src/main/java/campus/tech/kakao/map/Adapter/SearchDataAdapter.kt
@@ -1,11 +1,13 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.Adapter
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import campus.tech.kakao.map.Data.SearchData
+import campus.tech.kakao.map.R
+import campus.tech.kakao.map.ViewModel.RecentViewModel
 
 class SearchDataAdapter(
     private var items: List<SearchData>,
