+++ app/src/main/java/campus/tech/kakao/map/DBHelper/RecentDBHelper.kt
@@ -1,9 +1,10 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.DBHelper
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
 import android.util.Log
+import campus.tech.kakao.map.DataContract.RecentDataContract
 
 //version 2: DB에 time 컬럼을 추가함
 class RecentDBHelper(context: Context) : SQLiteOpenHelper(context, "RecentData.db", null, 2) {
