+++ app/src/main/java/campus/tech/kakao/map/Data/RecentSearchData.kt
@@ -1,3 +1,3 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.Data
 
 data class RecentSearchData(val name: String, val time: Long)
