+++ app/src/main/java/campus/tech/kakao/map/Data/SearchData.kt
@@ -1,3 +1,3 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.Data
 
 data class SearchData(val name: String, val category: String, val address: String)
